# contracts/openapi.yaml
...
openapi: 3.0.3
info:
  title: TaskFlow API
  version: 1.1.0
  description: A small multi-tenant task manager - login, logout, and CRUD over tasks.
servers:
  - url: http://127.0.0.1:8000
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Exchange credentials for a bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "401":
          description: Credentials rejected. Identical for unknown user and wrong password.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /me:
    get:
      tags: [auth]
      summary: The authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "401":
          description: Not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /tasks:
    get:
      tags: [tasks]
      summary: List the caller's tasks, newest first
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status_filter
          in: query
          required: false
          schema: { $ref: "#/components/schemas/TaskStatus" }
        - name: q
          in: query
          required: false
          description: Case-insensitive substring match on the title
          schema: { type: string }
      responses:
        "200":
          description: Matching tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TaskResponse" }
    post:
      tags: [tasks]
      summary: Create a task. Always starts in `todo`.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskCreateRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskResponse" }
        "422":
          description: A domain rule rejected the input
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /test/users:
    post:
      tags: [test-support]
      summary: Test-data factory. 404 unless ENABLE_TEST_ENDPOINTS=true.
      parameters:
        - { name: username, in: query, required: false, schema: { type: string } }
        - { name: password, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Created user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "404":
          description: Test endpoints are disabled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TaskStatus:
      type: string
      enum: [todo, in_progress, done]
    LoginRequest:
      type: object
      required: [username, password]
      additionalProperties: false
      properties:
        username: { type: string, minLength: 1, maxLength: 64 }
        password: { type: string, minLength: 1, maxLength: 256 }
    TokenResponse:
      type: object
      required: [access_token, token_type]
      properties:
        access_token: { type: string }
        token_type: { type: string, enum: [bearer] }
    UserResponse:
      type: object
      required: [id, username, tenant_id]
      properties:
        id: { type: string }
        username: { type: string }
        tenant_id:
          type: string
          description: Which customer this user belongs to.
...
    TaskCreateRequest:
      type: object
      required: [title]
      additionalProperties: false
      properties:
        title: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 2000, nullable: true }
...
    TaskResponse:
      type: object
      required: [id, owner_id, title, description, status, created_at, updated_at]
      properties:
        id: { type: string }
        owner_id: { type: string }
        title: { type: string }
        description: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ErrorResponse:
      type: object
      required: [code, detail]
      description: Every 4xx uses this shape. Branch on `code`, never on `detail`.
      properties:
        code: { type: string }
        detail: { type: string }